Be smarter about generating man pages
With the current rules, the man pages will be generated repeatedly if you
have xmlto installed. This is because make always thinks they are out of
date with respect to their prerequisite, do_xmlto_stage. They are changed
here to handle the multiple output files as described in the automake
manual.

distcheck has been confirmed to still work such that building from the
tarball does not require xmlto. On the other hand, if the user wants to
rebuild the man pages, they'll have to explicitly touch XI.xml.

It may be better to split the xml into per-page files so that xmlto only
generates one output at a time. A toplevel XI.xml can still pull together
all the pieces with XInclude.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>