Switch man pages to asciidoc.
The docbook xml is unreadable and unwritable. Switch everything to asciidoc,
at least thay we we can write man pages without losing sanity.

asciidoc seems to have some issues with __libmansuffix__, so instead we just
use libmansuffix (without underscores).
The current asciidoc version is buggy for multiple man targets (more than
one man page in a single file), so stick a big warning into configure that a
patch is required to the the right asciidoc->docbook conversion.

Many thanks to Dan Nicholson for the Makefile.am.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>