XQueryDeviceState: correct length of XValuatorState
The length is suppposed to be the complete class size, not just the struct
size. Since XValuatorState is followed by multiple valuator values the
size of the class is variable and dependent on the number of valuators.

The server assembles the valuator state last. This bug is unlikely to affect
clients as they should never go past the last class anyway.

Tested-by: Thomas Jaeger
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>