Handle unknown device classes.
If the server sends an unknown device class in response to an XIQueryDevice
call, no memory is allocated for these classes but we still write type
and sourceid as well as setting to->classes[i]. The latter causes multiple
classes to point to the same memory field.

Move the common code of assigning these three into the respective class type
handlers so to automatically skip any unknown classes.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>