_XIPassiveGrabDevice needs to set time value
When setting up a XIPassiveGrabDevice request, the time field is not
being set, leading to improper data being passed 'over the wire'.

Accept a time value into _XIPassiveGrabDevice and use it to set the time
field in the request.  Since the the functions calling
_XIPassiveGrabDevice are part of the API, and they do not accept time
values, they can just pass CurrentTime.

Signed-off-by: Jeff Smith <whydoubt@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>