Fix compile error that slipped through
Lesson for the future: make sure the private environment (e.g
convenience scripts) don't add changes where they are not expected. In
my case I'm building and installing Debian packages straigth away. And
one of the extra commits for Debian support had a change that had to go
to the libXi repository.

The development snapshot 1.7.99.1 has thus not been released (only a
commit with the version bump was pushed).

Signed-off-by: Povilas Kanapickas <povilas@radix.lt>